pitch-systems-api/pyproject.toml
[tool.poetry]
name = "pitch-system-api"
version = "0.1.0"
description = ""
authors = ["Aonrud <aonrud@fastmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.7"
djangorestframework = "^3.15.2"
mysqlclient = "^2.2.4"
drf-spectacular = "^0.27.2"

[tool.poetry.group.dev.dependencies]
django-stubs = "^5.0.2"
python-dotenv = "^1.0.1"
black = "^24.4.2"
djangorestframework-stubs = "^3.15.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
plugins = ["mypy_django_plugin.main"]

[tool.django-stubs]
django_settings_module = "pitch_systems_api.settings"
